Remove of the unnecessary view model. Add new method for token transaction view model.
